working on authorization checker

+++ Backend/src/main/java/cz/cvut/fit/genepi/presentationLayer/controller/PatientController.java
@@ -7,6 +7,7 @@ import cz.cvut.fit.genepi.dataLayer.entity.ExportParamsEntity;
 import cz.cvut.fit.genepi.dataLayer.entity.PatientEntity;
 import cz.cvut.fit.genepi.dataLayer.entity.RoleEntity;
 import cz.cvut.fit.genepi.dataLayer.entity.UserEntity;
+import cz.cvut.fit.genepi.util.AuthorizationChecker;
 import cz.cvut.fit.genepi.util.JSONEncoder;
 import cz.cvut.fit.genepi.util.LoggingService;
 import cz.cvut.fit.genepi.util.TimeConverter;
@@ -20,6 +21,7 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
@@ -89,7 +91,10 @@ public class PatientController {
      * @return the string
      */
     @RequestMapping(value = "/patient/create", method = RequestMethod.GET)
-    public String patientCreateGET(Locale locale, Model model) {
+    public String patientCreateGET(Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("doctors", roleService.getAllDoctors());
         model.addAttribute("patient", new PatientEntity());
         return "patient/createView";
@@ -106,7 +111,10 @@ public class PatientController {
     public String patientCreatePOST(
             @ModelAttribute("patient") @Valid PatientEntity patient,
             BindingResult result, @RequestParam("doctorId") Integer doctorId,
-            Locale locale, Model model) {
+            Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         if (result.hasErrors() || TimeConverter.compareDates(patient.getBirthday(), DateTime.now())) {
             model.addAttribute("doctors", roleService.getAllDoctors());
             return "patient/createView";
@@ -129,14 +137,20 @@ public class PatientController {
      */
     @RequestMapping(value = "/patientOverview", method = RequestMethod.POST)
     public String patientOverviewPOST(@RequestParam("id") int id,
-                                      Locale locale, Model model) {
+                                      Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("id");
         return "patient/overviewView";
     }
 
     @RequestMapping(value = "/patient/{patientId}/verify", method = RequestMethod.GET)
     public String patientVerifyGET(
-            @PathVariable("patientId") Integer patientId, Locale locale, Model model) {
+            @PathVariable("patientId") Integer patientId, Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
 
         PatientDisplayVO patient = patientService.getPatientDisplayByIdWithDoctor(patientId);
         model.addAttribute("patient", patient);
@@ -154,7 +168,10 @@ public class PatientController {
      */
     @RequestMapping(value = "/patient/{patientId}/overview", method = RequestMethod.GET)
     public String patientOverviewGET(
-            @PathVariable("patientId") Integer patientId, Locale locale, Model model) {
+            @PathVariable("patientId") Integer patientId, Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
 
         PatientDisplayVO patient = patientService.getPatientDisplayByIdWithAllLists(patientId);
 
@@ -172,8 +189,10 @@ public class PatientController {
     @RequestMapping(value = "/patient/{patientId}/verify", method = RequestMethod.POST)
     public String patientVerifyPOST(
             @PathVariable("patientId") Integer patientId,
-            @RequestParam("verification") int verification, Locale locale, Model model) {
-
+            @RequestParam("verification") int verification, Locale locale, Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         if (verification == 1) {
             patientService.verifyPatient(patientId);
         } else if (verification == 2) {
@@ -214,7 +233,7 @@ public class PatientController {
      */
     @RequestMapping(value = "/patient/list", method = RequestMethod.GET)
     public String patientsListGET(Locale locale, Model model,
-                                  @RequestParam("maxResults") int maxResults /*,
+                                  @RequestParam("maxResults") int maxResults , HttpServletRequest request/*,
                                   @RequestParam("pageNumber") int pageNumber*/) {
       /*  model.addAttribute("patientList", patientService.findAllWithPagination(PatientEntity.class, maxResults, pageNumber));
         model.addAttribute("maxResults", maxResults);
@@ -222,6 +241,9 @@ public class PatientController {
         // hotfix - there will be some method for getting the count yet
         model.addAttribute("countOfPatients",
                 patientService.getCount(PatientEntity.class));*/
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("maxResult", maxResults);
         return "patient/listView";
     }
@@ -240,8 +262,10 @@ public class PatientController {
     String patientsListSearchGET(Locale locale, Model model,
                                  @RequestParam("maxResults") int maxResults,
                                  @RequestParam("pageNumber") int pageNumber,
-                                 @RequestParam("search") String searchString) {
-
+                                 @RequestParam("search") String searchString, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         List<String> searchParams = new ArrayList<>();
         searchParams.add("contact.firstName");
         searchParams.add("contact.lastName");
@@ -269,7 +293,10 @@ public class PatientController {
      */
     @RequestMapping(value = "/patient/{patientID}/delete", method = RequestMethod.GET)
     public String patientDeleteGET(Locale locale, Model model,
-                                   @PathVariable("patientID") Integer patientID) {
+                                   @PathVariable("patientID") Integer patientID, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         patientService.delete(patientService.findByID(PatientEntity.class,
                 patientID));
         return "redirect:/patient/list";
@@ -284,7 +311,10 @@ public class PatientController {
      */
     @RequestMapping(value = "/patient/{patientId}/edit", method = RequestMethod.GET)
     public String patientEditGET(Locale locale, Model model,
-                                 @PathVariable("patientId") Integer patientId) {
+                                 @PathVariable("patientId") Integer patientId, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         model.addAttribute("patient",
                 patientService.getPatientByIdWithDoctor(patientId));
         model.addAttribute("doctors", roleService.getAllDoctors());
@@ -302,7 +332,10 @@ public class PatientController {
     @RequestMapping(value = "/patient/{patientId}/edit", method = RequestMethod.POST)
     public String patientEditPOST(Locale locale, Model model,
                                   @Valid @ModelAttribute("patient") PatientEntity patient,
-                                  BindingResult result) {
+                                  BindingResult result, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
 
         if (result.hasErrors()) {
             model.addAttribute("doctors", roleService.getAllDoctors());
@@ -335,8 +368,10 @@ public class PatientController {
 
     @RequestMapping(value = "/patient/{patientID}/export", method = RequestMethod.GET)
     public String patientExportGET(Locale locale, Model model,
-                                   @PathVariable("patientID") Integer patientID) {
-
+                                   @PathVariable("patientID") Integer patientID, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         Authentication auth = SecurityContextHolder.getContext()
                 .getAuthentication();
         UserEntity user = userService.findUserByUsername(auth.getName());
@@ -381,7 +416,10 @@ public class PatientController {
             @RequestParam("patient") Integer[] patientID,
             //@RequestParam("toTable") boolean toTable,
             @RequestParam("exportType") String exportType, Locale locale, boolean anonymize,
-            Model model) {
+            Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         logger.setLogger(PatientController.class);
 
         Authentication auth = SecurityContextHolder.getContext()
@@ -611,8 +649,10 @@ public class PatientController {
             @RequestParam("patient") Integer[] patient,
             @RequestParam("isGeneric") boolean isGeneric,
             @RequestParam("exportName") String exportName, Locale locale,
-            Model model) {
-
+            Model model, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         exportParams.setName(exportName);
         Authentication auth = SecurityContextHolder.getContext()
                 .getAuthentication();
@@ -628,8 +668,10 @@ public class PatientController {
     @RequestMapping(value = "/patient/export/load", method = RequestMethod.POST)
     public String patientExportLoadPOST(Model model, Locale locale,
                                         @RequestParam("patient") Integer[] patientID,
-                                        @RequestParam("exportId") Integer exportID) {
-
+                                        @RequestParam("exportId") Integer exportID, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         if (logger.getLogger() == null)
             logger.setLogger(PatientController.class);
 
@@ -680,7 +722,10 @@ public class PatientController {
     @RequestMapping(value = "/patient/export/delete", method = RequestMethod.POST)
     public String patientExportDeletePOST(
             @RequestParam("patient") Integer[] patientId,
-            @RequestParam("exportId") Integer exportId) {
+            @RequestParam("exportId") Integer exportId, HttpServletRequest request) {
+        if (!AuthorizationChecker.checkAuthoritaion(request)) {
+            return "deniedView";
+        }
         exportParamsService.delete(exportParamsService.findByID(
                 ExportParamsEntity.class, exportId));
         return "redirect:/patient/" + patientId[0] + "/export";
